Updated stable

+++ code/stable/projectile/Projectile_S_NoL_U_U_V/src/cpp/Calculations.cpp
@@ -5,19 +5,19 @@
 
 using std::string;
 
-double func_t_flight(double v_launch, double theta, double g_vect) {
+float func_t_flight(float v_launch, float theta, float g_vect) {
     return 2 * v_launch * sin(theta) / g_vect;
 }
 
-double func_p_land(double v_launch, double theta, double g_vect) {
+float func_p_land(float v_launch, float theta, float g_vect) {
     return 2 * pow(v_launch, 2) * sin(theta) * cos(theta) / g_vect;
 }
 
-double func_d_offset(double p_target, double p_land) {
+float func_d_offset(float p_target, float p_land) {
     return p_land - p_target;
 }
 
-string func_s(double p_target, double epsilon, double d_offset) {
+string func_s(float p_target, float epsilon, float d_offset) {
     if (fabs(d_offset / p_target) < epsilon) {
         return "The target was hit.";
     }
