+++ code/stable/projectile/Projectile_S_NoL_U_U_V/src/cpp/InputConstraints.hpp
@@ -14,6 +14,6 @@
     \param theta launch angle: the angle between the launcher and a straight line from the launcher to the target (rad)
     \param p_target target position: the distance from the launcher to the target (m)
 */
-void input_constraints(double v_launch, double theta, double p_target);
+void input_constraints(float v_launch, float theta, float p_target);
 
 #endif
