+++ code/stable/projectile/Projectile_S_NoL_U_U_V/src/cpp/OutputFormat.hpp
@@ -14,6 +14,6 @@ using std::string;
     \param s output message as a string
     \param d_offset distance between the target position and the landing position: the offset between the target position and the landing position (m)
 */
-void write_output(string s, double d_offset);
+void write_output(string s, float d_offset);
 
 #endif
