+++ code/stable/projectile/Projectile_S_NoL_U_U_V/src/csharp/OutputFormat.cs
@@ -11,7 +11,7 @@ public class OutputFormat {
         \param s output message as a string
         \param d_offset distance between the target position and the landing position: the offset between the target position and the landing position (m)
     */
-    public static void write_output(string s, double d_offset) {
+    public static void write_output(string s, float d_offset) {
         StreamWriter outputfile;
         outputfile = new StreamWriter("output.txt", false);
         outputfile.Write("s = ");
