+++ code/stable/projectile/Projectile_S_NoL_U_U_V/src/java/Projectile/Control.java
@@ -14,19 +14,19 @@ public class Control {
     */
     public static void main(String[] args) throws FileNotFoundException, IOException {
         String filename = args[0];
-        double v_launch;
-        double theta;
-        double p_target;
-        double g_vect = 9.8;
-        double epsilon = 2.0e-2;
+        float v_launch;
+        float theta;
+        float p_target;
+        float g_vect = 9.8f;
+        float epsilon = 2.0e-2f;
         Object[] outputs = InputFormat.get_input(filename);
-        v_launch = (double)(outputs[0]);
-        theta = (double)(outputs[1]);
-        p_target = (double)(outputs[2]);
+        v_launch = (float)(outputs[0]);
+        theta = (float)(outputs[1]);
+        p_target = (float)(outputs[2]);
         InputConstraints.input_constraints(v_launch, theta, p_target);
-        double t_flight = Calculations.func_t_flight(v_launch, theta, g_vect);
-        double p_land = Calculations.func_p_land(v_launch, theta, g_vect);
-        double d_offset = Calculations.func_d_offset(p_target, p_land);
+        float t_flight = Calculations.func_t_flight(v_launch, theta, g_vect);
+        float p_land = Calculations.func_p_land(v_launch, theta, g_vect);
+        float d_offset = Calculations.func_d_offset(p_target, p_land);
         String s = Calculations.func_s(p_target, epsilon, d_offset);
         OutputFormat.write_output(s, d_offset);
     }
